azureresourcesdks/requirements.txt
azure-mgmt-resource>=18.0.0
azure-identity>=1.5.0
azure-mgmt-resource
azure-mgmt-compute
azure-mgmt-network
azure-identity
azure-mgmt-rdbms
mysql
mysql-connector